theme-sync/requirements.txt
watchdog==3.0.0
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
pystache==0.6.5

